clients/fastapi/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
protobuf>=4.21.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
aiohttp>=3.8.0
